Created visual task list UI elements & system to give the player some feedback on completing tasks + Created basic health bar + Tweaked task system slightly to let me test tasks with a single input sequence length

+++ U.BarrierToEntry/Assets/DSmyth/Scripts/StatesModule/GameStates.cs
@@ -8,8 +8,10 @@ namespace DSmyth.StatesModule
         public static Action OnGameplayInititalized;
         public static Action OnGameOver;
 
-
-        public static Action<int> OnHealthChanged;
+        /// <summary>
+        /// Passes through the Current Health percentage whenever the Current Health value changes.
+        /// </summary>
+        public static Action<float> OnHealthChanged;
         public static Action<int> OnScoreChanged;
     }
 }
